[update] split dataset.py into multiple files to enhance code readability
